requirements.txt
langchain 
langchain-openai
python-dotenv
langchain_community
pypdf
bs4
arxiv
fitz
frontend
wikipedia
langchain-text-splitters
langchain_openai
chromadb
langchain_ollama
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
google-generativeai
groq
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
streamlit
pymysql
validators
youtube_transcript_api
pytube
yt-dlp
gradio
langchain_nvidia_ai_endpoints
crewai
crewai_tools
